refactor(trie): move `Trie#_findValueNodes` to `TrieReadStream`
